requirements.txt
# Python依赖包
python-telegram-bot>=20.0
httpx>=0.25.0
playwright>=1.40.0
prettytable>=3.10.0
tqdm>=4.66.0
colorama

# 网盘爬虫系统新增依赖
beautifulsoup4>=4.12.0
pandas>=2.1.0
numpy>=1.25.0
fake-useragent>=1.4.0
loguru>=0.7.0
python-dotenv>=1.0.0
aiofiles>=23.2.0

# 可选依赖
requests>=2.31.0
aiohttp>=3.8.0
lxml>=4.9.0